+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/DispatcherServlet.java
@@ -1,12 +1,9 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
-import com.interface21.webmvc.servlet.view.JspView;
-import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
@@ -19,18 +16,14 @@ public class DispatcherServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
-    private ManualHandlerMapping manualHandlerMapping;
     private AnnotationHandlerMapping annotationHandlerMapping;
 
-    public DispatcherServlet(ManualHandlerMapping manualHandlerMapping,
-                             AnnotationHandlerMapping annotationHandlerMapping) {
-        this.manualHandlerMapping = manualHandlerMapping;
+    public DispatcherServlet(AnnotationHandlerMapping annotationHandlerMapping) {
         this.annotationHandlerMapping = annotationHandlerMapping;
     }
 
     @Override
     public void init() {
-        manualHandlerMapping.initialize();
         annotationHandlerMapping.initialize();
     }
 
@@ -42,9 +35,7 @@ public class DispatcherServlet extends HttpServlet {
         HandlerExecution handlerExecution = (HandlerExecution) annotationHandlerMapping.getHandler(request);
         if (handlerExecution != null) {
             handleWithHandlerExecution(handlerExecution, request, response);
-            return;
         }
-        handelWithManualHandlerMapping(request, response);
     }
 
     private void handleWithHandlerExecution(HandlerExecution handlerExecution, HttpServletRequest request,
@@ -57,29 +48,5 @@ public class DispatcherServlet extends HttpServlet {
             log.error("Exception : {}", e.getMessage(), e);
             throw new ServletException(e.getMessage());
         }
-    }
-
-    private void handelWithManualHandlerMapping(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException {
-        String requestURI = request.getRequestURI();
-        try {
-            Controller controller = manualHandlerMapping.getHandler(requestURI);
-            String viewName = controller.execute(request, response);
-            move(viewName, request, response);
-        } catch (Throwable e) {
-            log.error("Exception : {}", e.getMessage(), e);
-            throw new ServletException(e.getMessage());
-        }
-    }
-
-    private void move(String viewName, HttpServletRequest request, HttpServletResponse response)
-            throws Exception {
-        if (viewName.startsWith(JspView.REDIRECT_PREFIX)) {
-            response.sendRedirect(viewName.substring(JspView.REDIRECT_PREFIX.length()));
-            return;
-        }
-
-        RequestDispatcher requestDispatcher = request.getRequestDispatcher(viewName);
-        requestDispatcher.forward(request, response);
     }
 }
